data

+++ start.sh
@@ -2,7 +2,6 @@
 
 # Iniciar la aplicación FastAPI
 uvicorn api.main:app --host 0.0.0.0 --port 8000 &
-
 # Guardar el PID del proceso de FastAPI
 UVICORN_PID=$!
 
@@ -15,17 +14,24 @@ else
     echo "El bot de Telegram ya está en ejecución."
 fi
 
-# Iniciarla API
+# Iniciar la API (si es diferente del bot)
 python api/get.py &
-# Guardar el PID del proceso de la api
+# Guardar el PID del proceso de la API
 API_PID=$!
 
 # Función para manejar la terminación
 cleanup() {
     echo "Shutting down..."
-    kill $UVICORN_PID
-    kill $BOT_PID
-    kill $API_PID
+    # Asegurarse de que solo se intenten matar procesos en ejecución
+    if kill -0 $UVICORN_PID >/dev/null 2>&1; then
+        kill $UVICORN_PID
+    fi
+    if [[ -n "$BOT_PID" ]] && kill -0 $BOT_PID >/dev/null 2>&1; then
+        kill $BOT_PID
+    fi
+    if [[ -n "$API_PID" ]] && kill -0 $API_PID >/dev/null 2>&1; then
+        kill $API_PID
+    fi
     exit
 }
